refactor: shop -> workspace 폴더명 변경
